feat: split grpc server into static and dynamic varriant
